Login Page
Setting up login page for the website. There will be two text boxes where the user will enter their email or password. The buttons give the user the option of signing in or signing up (This functionality of the buttons working for the user to sign in/sign up will be created later). Focus is on getting the page setup. Code used is c# on visual studio.