SQL Connection
Created a SQL database in SQL Server Management Studio. When the user enters in their credentials in the username and password text box, they will then click the sign in button. If the credentials match what's currently in the database, the user will receive a login success button. If the user leaves the text boxes null or enters the incorrect information, an error message will appear "Error: Please check your login credentials."
Currently, I have not started the functionality for the sign up button. This will be updated in the future.