added dropdown box
unfortunately, I lost all the data to the previous project so I had to recreate it. To pick up where I left off, I added three new label links titled Browse Routines, Fitness Progress, and Workout Plans. When the user logs in, the browse routines window will be the default window. I added a dropdown list where the user can select a muscle group they will like to focus on from the dropdown list. Once value is selected, it will be compared to the value within the muscle_group column within  browseRoutines table. Once located, it will then display a list of exercise names that tie to the value that was selected, along with the youtube video to learn how to do the movement.