Fitness Progress Feature Completed
Finished creating fitness progress feature. The feature takes in the users goal weight and current weight and provides a message in the listBox based on the user's input. The database CurrentWeight and GoalWeight also get updated based on the user input. The email used in the login form is used as an identifier for the user.