Workout Plan feature
Added feature where the user will be able to generate a workout plan based on the day of the week. The date will be provided along with the date plan and the duration of the workout plan. The feature works, but currently experiencing trouble starting the display message on a new line as it gets cut off in the listbox.